Reinstante GNU Readline as an option...
Option --readline is now a choice: GNU, Prompt, or None

Prompt is prompt-toolkit and is the default. However in some situations
where the terminal is limited we sometimes need GNU or none.
