+++ mathicsscript/termshell.py
@@ -95,12 +95,11 @@ class ShellEscapeException(Exception):
         self.line = line
 
 
-class TerminalShell(MathicsLineFeeder):
+class TerminalShellCommon(MathicsLineFeeder):
     def __init__(
         self,
         definitions,
         style: str,
-        want_readline: bool,
         want_completion: bool,
         use_unicode: bool,
         prompt: bool,
@@ -109,20 +108,208 @@ class TerminalShell(MathicsLineFeeder):
         self.input_encoding = locale.getpreferredencoding()
         self.lineno = 0
         self.terminal_formatter = None
+        self.prompt = prompt
+
+        colorama_init()
+        if style == "None":
+            self.terminal_formatter = None
+            self.incolors = self.outcolors = ["", "", "", ""]
+        else:
+            # self.incolors = ["\033[34m", "\033[1m", "\033[22m", "\033[39m"]
+            self.incolors = ["\033[32m", "\033[1m", "\033[22m", "\033[39m"]
+            self.outcolors = ["\033[31m", "\033[1m", "\033[22m", "\033[39m"]
+            if style is not None and not is_pygments_style(style):
+                style = None
+
+            if style is None:
+                dark_background = is_dark_background()
+                if dark_background:
+                    style = "inkpot"
+                else:
+                    style = "colorful"
+            try:
+                self.terminal_formatter = Terminal256Formatter(style=style)
+            except ClassNotFound:
+                print(
+                    "Pygments style name '%s' not found; No pygments style set" % style
+                )
+
+        self.pygments_style = style
+        self.definitions = definitions
+        self.definitions.set_ownvalue(
+            "Settings`$PygmentsShowTokens", from_python(False)
+        )
+        self.definitions.set_ownvalue("Settings`$PygmentsStyle", from_python(style))
+        self.definitions.set_ownvalue("Settings`$UseUnicode", from_python(use_unicode))
+        self.definitions.set_ownvalue(
+            "Settings`PygmentsStylesAvailable", from_python(ALL_PYGMENTS_STYLES)
+        )
+        self.definitions.add_message(
+            "Settings`PygmentsStylesAvailable",
+            Rule(
+                Expression(
+                    "System`MessageName",
+                    Symbol("Settings`PygmentsStylesAvailable"),
+                    from_python("usage"),
+                ),
+                from_python(
+                    "A list of Pygments style that are valid in Settings`$PygmentsStyle."
+                ),
+            ),
+        )
+        self.definitions.set_attribute(
+            "Settings`PygmentsStylesAvailable", "System`Protected"
+        )
+        self.definitions.set_attribute(
+            "Settings`PygmentsStylesAvailable", "System`Locked"
+        )
+        self.definitions.set_attribute("Settings`$UseUnicode", "System`Locked")
+
+    def change_pygments_style(self, style: str):
+        if style == self.pygments_style:
+            return False
+        if is_pygments_style(style):
+            self.terminal_formatter = Terminal256Formatter(style=style)
+            self.pygments_style = style
+            return True
+        else:
+            print("Pygments style not changed")
+            return False
+
+    def get_last_line_number(self):
+        return self.definitions.get_line_no()
+
+    def get_in_prompt(self):
+        next_line_number = self.get_last_line_number() + 1
+        if self.lineno > 0:
+            return " " * len(f"In[{next_line_number}]:= ")
+        else:
+            return "{1}In[{2}{0}{3}]:= {4}".format(next_line_number, *self.incolors)
+            # if have_full_readline:
+            #     return "{1}In[{2}{0}{3}]:= {4}".format(next_line_number, *self.incolors)
+            # else:
+            #     return f"In[{next_line_number}]:= "
+
+    def get_out_prompt(self):
+        line_number = self.get_last_line_number()
+        return "{1}Out[{2}{0}{3}]= {4}".format(line_number, *self.outcolors)
+
+    def to_output(self, text):
+        line_number = self.get_last_line_number()
+        newline = "\n" + " " * len("Out[{0}]= ".format(line_number))
+        return newline.join(text.splitlines())
+
+    def out_callback(self, out):
+        print(self.to_output(str(out)))
+
+    def read_line(self, prompt):
+        if self.using_readline:
+            line = self.rl_read_line(prompt)
+        else:
+            line = input(prompt)
+        if line.startswith("!") and self.lineno == 0:
+            raise ShellEscapeException(line)
+        return replace_unicode_with_wl(line)
+
+    def print_result(
+        self, result, prompt: bool, output_style="", strict_wl_output=False
+    ):
+        if result is None:
+            # FIXME decide what to do here
+            return
+
+        last_eval = result.last_eval
+
+        if last_eval is not None:
+            try:
+                eval_type = last_eval.get_head_name()
+            except:
+                print(sys.exc_info()[1])
+                return
+
+            out_str = str(result.result)
+            use_highlight = True
+            if eval_type == "System`String":
+                if strict_wl_output:  # exact-wl-compatibility
+                    out_str = (
+                        format(
+                            [(MToken.STRING, out_str.rstrip())], self.terminal_formatter
+                        )
+                        + "\n"
+                    )
+                    use_highlight = False
+                else:
+                    out_str = '"' + out_str.replace('"', r"\"") + '"'
+            if eval_type == "System`Graph":
+                out_str = "-Graph-"
+            elif self.terminal_formatter:  # pygmentize
+                show_pygments_tokens = self.definitions.get_ownvalue(
+                    "Settings`$PygmentsShowTokens"
+                ).replace.to_python()
+                pygments_style = self.definitions.get_ownvalue(
+                    "Settings`$PygmentsStyle"
+                ).replace.get_string_value()
+                if pygments_style != self.pygments_style:
+                    if not self.change_pygments_style(pygments_style):
+                        self.definitions.set_ownvalue(
+                            "Settings`$PygmentsStyle", String(self.pygments_style)
+                        )
+
+                if show_pygments_tokens:
+                    print(list(lex(out_str, mma_lexer)))
+                if use_highlight:
+                    out_str = highlight(out_str, mma_lexer, self.terminal_formatter)
+            output = self.to_output(out_str)
+            if output_style == "text" or not prompt:
+                print(output)
+            else:
+                print(self.get_out_prompt() + output + "\n")
+
+    def rl_read_line(self, prompt):
+        # Wrap ANSI color sequences in \001 and \002, so readline
+        # knows that they're nonprinting.
+        prompt = self.ansi_color_re.sub(lambda m: "\001" + m.group(0) + "\002", prompt)
+
+        return input(prompt)
+
+    def reset_lineno(self):
+        self.lineno = 0
+
+    def feed(self):
+        prompt_str = self.get_in_prompt() if self.prompt else ""
+        result = self.read_line(prompt_str) + "\n"
+        if result == "\n":
+            return ""  # end of input
+        self.lineno += 1
+        return result
+
+    def empty(self):
+        return False
+
+
+class TerminalShellPromptToolKit(TerminalShellCommon):
+    def __init__(
+        self,
+        definitions,
+        style: str,
+        want_completion: bool,
+        use_unicode: bool,
+        prompt: bool,
+    ):
+        super(TerminalShellCommon, self).__init__("<stdin>")
+        self.input_encoding = locale.getpreferredencoding()
+        self.lineno = 0
+        self.terminal_formatter = None
         self.mma_pygments_lexer = PygmentsLexer(MathematicaLexer)
         self.prompt = prompt
 
-        if want_readline:
-            self.session = PromptSession(history=FileHistory(HISTFILE))
-        else:
-            self.session = None
+        self.session = PromptSession(history=FileHistory(HISTFILE))
 
         # Try importing readline to enable arrow keys support etc.
         self.using_readline = False
         self.history_length = definitions.get_config_value("$HistoryLength", HISTSIZE)
-        if want_readline:
-            self.using_readline = sys.stdin.isatty() and sys.stdout.isatty()
-            self.ansi_color_re = re.compile("\033\\[[0-9;]+m")
+        self.using_readline = sys.stdin.isatty() and sys.stdout.isatty()
+        self.ansi_color_re = re.compile("\033\\[[0-9;]+m")
 
         colorama_init()
         if style == "None":
@@ -199,20 +386,6 @@ class TerminalShell(MathicsLineFeeder):
             f" mathicsscript: {__version__}, Style: {self.pygments_style}, Mode: [F4] {edit_mode}"
         )
 
-    def change_pygments_style(self, style: str):
-        if style == self.pygments_style:
-            return False
-        if is_pygments_style(style):
-            self.terminal_formatter = Terminal256Formatter(style=style)
-            self.pygments_style = style
-            return True
-        else:
-            print("Pygments style not changed")
-            return False
-
-    def get_last_line_number(self):
-        return self.definitions.get_line_no()
-
     def get_in_prompt(self):
         next_line_number = self.get_last_line_number() + 1
         if self.lineno > 0:
@@ -222,40 +395,7 @@ class TerminalShell(MathicsLineFeeder):
 
     def get_out_prompt(self):
         line_number = self.get_last_line_number()
-        return f"{1}Out[{2}{0}{3}]= {4}".format(line_number, *self.outcolors)
-
-    def get_out_prompt_toolkit(self):
-        line_number = self.get_last_line_number()
         return HTML(f"<ansigreen>Out[<b>{line_number}</b>]:=</ansigreen> ")
-
-    def to_output(self, text):
-        line_number = self.get_last_line_number()
-        newline = "\n" + " " * len("Out[{0}]= ".format(line_number))
-        return newline.join(text.splitlines())
-
-    def out_callback(self, out):
-        print(self.to_output(str(out)))
-
-    def read_line(self, prompt):
-        if self.using_readline and self.session:
-
-            # FIXME set and update inside self.
-            style = style_from_pygments_cls(get_style_by_name(self.pygments_style))
-
-            line = self.session.prompt(
-                prompt,
-                bottom_toolbar=self.bottom_toolbar,
-                completer=self.completer,
-                key_bindings=bindings,
-                lexer=self.mma_pygments_lexer,
-                style=style,
-            )
-            # line = self.rl_read_line(prompt)
-        else:
-            line = input(prompt)
-        if line.startswith("!") and self.lineno == 0:
-            raise ShellEscapeException(line)
-        return replace_unicode_with_wl(line)
 
     def print_result(
         self, result, prompt: bool, output_style="", strict_wl_output=False
@@ -309,28 +449,24 @@ class TerminalShell(MathicsLineFeeder):
             if output_style == "text" or not prompt:
                 print(output)
             elif self.session:
-                print_formatted_text(self.get_out_prompt_toolkit(), end="")
+                print_formatted_text(self.get_out_prompt(), end="")
                 print(output + "\n")
             else:
                 print(self.get_out_prompt() + output + "\n")
 
-    def rl_read_line(self, prompt):
-        # Wrap ANSI color sequences in \001 and \002, so readline
-        # knows that they're nonprinting.
-        prompt = self.ansi_color_re.sub(lambda m: "\001" + m.group(0) + "\002", prompt)
+    def read_line(self, prompt):
+        # FIXME set and update inside self.
+        style = style_from_pygments_cls(get_style_by_name(self.pygments_style))
 
-        return input(prompt)
-
-    def reset_lineno(self):
-        self.lineno = 0
-
-    def feed(self):
-        prompt_str = self.get_in_prompt() if self.prompt else ""
-        result = self.read_line(prompt_str) + "\n"
-        if result == "\n":
-            return ""  # end of input
-        self.lineno += 1
-        return result
-
-    def empty(self):
-        return False
+        line = self.session.prompt(
+            prompt,
+            bottom_toolbar=self.bottom_toolbar,
+            completer=self.completer,
+            key_bindings=bindings,
+            lexer=self.mma_pygments_lexer,
+            style=style,
+        )
+        # line = self.rl_read_line(prompt)
+        if line.startswith("!") and self.lineno == 0:
+            raise ShellEscapeException(line)
+        return replace_unicode_with_wl(line)
